Added sleeping.
Use --sleep YTID TIME to mark a video as "sleeping" which means it will not be checked until the sleep timer is up. This is useful for live or premiering videos to avoid repeatedly checking them in the mean time.
Use --sleep YTID to get the current time (if set).
Use --unsleep YTID to mark it is no longer sleeping.

TIME can be an absolute YYYY-MM-DD HH:MM:SS time, or relative with d+N for days, h+N for hours, m+N for minutes, s+N for seconds from the current time.
All times are in UTC.